requirements.txt
pandas
numpy
langchain
langgraph
openai
faiss-cpu
tavily-python
wikipedia
pyyaml
requests
python-dotenv
mcp
fastmcp
fastapi
uvicorn
presidio-analyzer
presidio-anonymizer
spacy
tqdm
PyPDF2
langchain-community
langchain-openai
sentence-transformers
groq
